Mask Slide Handler
Added:

LinerInterpolation - this interpolates how much skin to reveal based of the size of the object.

MoveMask - In charge of the logic that address the layer you want to reveal and via precentage of the overall size of the yet to be drink.

SlideMask - public incapulation of Move mask, client will enter the amount of drink they want to add in precentage and the intended layer this behind the scene calls Movemask and addes the desired percentage to the currentprecentage

Debug region - staying as of right now for debug purpose, * * * You will need to public / serialize currentPrecenatge and currentMaskInOrder